+++ src/runtime/virtcontainers/acrn_arch_base_test.go
@@ -3,6 +3,9 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+//go:build linux
+// +build linux
+
 package virtcontainers
 
 import (
