+++ src/runtime/virtcontainers/hypervisor_test.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"testing"
 
-	ktu "github.com/kata-containers/kata-containers/src/runtime/pkg/katatestutils"
 	"github.com/kata-containers/kata-containers/src/runtime/virtcontainers/types"
 	"github.com/stretchr/testify/assert"
 )
@@ -297,22 +296,6 @@ func genericTestRunningOnVMM(t *testing.T, data []testNestedVMMData) {
 
 		assert.Equal(running, d.expected)
 	}
-}
-
-func TestGenerateVMSocket(t *testing.T) {
-	assert := assert.New(t)
-
-	if tc.NotValid(ktu.NeedRoot()) {
-		t.Skip(testDisabledAsNonRoot)
-	}
-	s, err := generateVMSocket("a", "")
-	assert.NoError(err)
-	vsock, ok := s.(types.VSock)
-	assert.True(ok)
-	defer assert.NoError(vsock.VhostFd.Close())
-	assert.NotZero(vsock.VhostFd)
-	assert.NotZero(vsock.ContextID)
-	assert.NotZero(vsock.Port)
 }
 
 func TestAssetPath(t *testing.T) {
