+++ src/runtime/virtcontainers/vm_test.go
@@ -11,7 +11,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/kata-containers/kata-containers/src/runtime/virtcontainers/utils"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -84,45 +83,4 @@ func TestNewVM(t *testing.T) {
 	config.HypervisorConfig.DevicesStatePath = testDir
 	_, err = NewVM(ctx, config)
 	assert.Nil(err)
-}
-
-func TestVMConfigValid(t *testing.T) {
-	assert := assert.New(t)
-
-	config := VMConfig{}
-
-	err := config.Valid()
-	assert.Error(err)
-
-	testDir := t.TempDir()
-
-	config.HypervisorConfig = HypervisorConfig{
-		KernelPath: testDir,
-		InitrdPath: testDir,
-	}
-	err = config.Valid()
-	assert.Nil(err)
-}
-
-func TestVMConfigGrpc(t *testing.T) {
-	assert := assert.New(t)
-	config := VMConfig{
-		HypervisorType:   QemuHypervisor,
-		HypervisorConfig: newQemuConfig(),
-		AgentConfig: KataAgentConfig{
-			LongLiveConn:       true,
-			Debug:              false,
-			Trace:              false,
-			EnableDebugConsole: false,
-			ContainerPipeSize:  0,
-			KernelModules:      []string{}},
-	}
-
-	p, err := config.ToGrpc()
-	assert.Nil(err)
-
-	config2, err := GrpcToVMConfig(p)
-	assert.Nil(err)
-
-	assert.True(utils.DeepCompare(config, *config2))
 }
